+++ app/core/+mag/+app/+internal/removeNonFiniteData.m
@@ -1,0 +1,19 @@
+function data = removeNonFiniteData(data)
+% REMOVENONFINITEDATA Remove non-finite data from table or timetable.
+
+    arguments (Input)
+        data tabular
+    end
+
+    arguments (Output)
+        data tabular
+    end
+
+    locNonFinite = ~isfinite(data.Variables);
+
+    if any(locNonFinite, "all")
+
+        warning("mag:app:nonFiniteData", "Removing non-finite data.");
+        data(any(locNonFinite, 2), :) = [];
+    end
+end