+++ resources/release-notes.md
@@ -5,3 +5,4 @@
 ## App
 
 - (HelioSwarm, IMAP) Add support for Signal Analyzer and Wavelet Analyzer
+- (Bartington, HelioSwarm, IMAP) Add support for non-finite signals in Signal Analyzer and Wavelet Analyzer